refactor: move NewTestServer to wstest sub-package
1.Test infra polluted production binary → new wstest/ sub-package
2.Moved test from testing_test.go → wstest/server_test.go
3.Closes #29
