+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+### Changed
+
+- **BREAKING**: `NewTestServer` moved from the main `wspulse` package to `server/wstest` sub-package. Import path changes from `wspulse.NewTestServer(...)` to `wstest.NewTestServer(...)`. This removes `net/http/httptest` and `testing` from the production import graph.
+
+### Removed
+
+- `NewTestServer` from the main package (use `github.com/wspulse/server/wstest` instead)
+
 ---
 
 ## [0.7.0] - 2026-04-08
